JodoroidApplicationからHTTP関連のファクトリメソッドを切り出す
