+++ app/src/main/java/com/pepabo/jodo/jodoroid/JodoroidApplication.java
@@ -8,12 +8,11 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.Uri;
-import android.support.annotation.NonNull;
 
 import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.squareup.okhttp.Cache;
+import com.pepabo.jodo.jodoroid.modules.HttpModule;
 import com.squareup.picasso.OkHttpDownloader;
 import com.squareup.picasso.Picasso;
 
@@ -21,19 +20,9 @@ import com.squareup.picasso.Picasso;
 import com.pepabo.jodo.jodoroid.models.APIService;
 import com.squareup.okhttp.OkHttpClient;
 
-import java.io.File;
-import java.net.CookieHandler;
-import java.net.CookieManager;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
 import java.util.Date;
-
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSession;
-import javax.net.ssl.X509TrustManager;
 
 import retrofit.RequestInterceptor;
 import retrofit.RestAdapter;
@@ -43,7 +32,6 @@ import retrofit.converter.GsonConverter;
 
 public class JodoroidApplication extends Application implements OnAccountsUpdateListener {
     public static final String ENDPOINT = "https://157.7.190.186/api/";
-    private static final String CACHE_DIR_NAME = "http-cache";
 
     public static final String ACTION_LOGGED_OUT = "com.pepabo.jodo.jodoroid.LOGGED_OUT";
 
@@ -55,7 +43,7 @@ public class JodoroidApplication extends Application implements OnAccountsUpdate
         super.onCreate();
 
         try {
-            final OkHttpClient httpClient = getOkHttpClient(this);
+            final OkHttpClient httpClient = HttpModule.provideHttpClient(this);
             mPicasso = createPicasso(this, httpClient);
             mService = createAPIService(this, httpClient);
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
@@ -81,10 +69,6 @@ public class JodoroidApplication extends Application implements OnAccountsUpdate
                 .build();
     }
 
-    private static CookieHandler createCookieHandler() {
-        return new CookieManager(); // TODO: Implement persistent cookie jar
-    }
-
     private static Gson createGson() {
         return new GsonBuilder()
                 .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
@@ -104,46 +88,6 @@ public class JodoroidApplication extends Application implements OnAccountsUpdate
                 .build();
 
         return adapter.create(APIService.class);
-    }
-
-    @NonNull
-    private static OkHttpClient getOkHttpClient(Context context) throws NoSuchAlgorithmException, KeyManagementException {
-        final SSLContext sslContext = SSLContext.getInstance("TLS");
-        sslContext.init(null, new X509TrustManager[]{new TrustEveryoneX509TrustManager()}, null);
-
-        final File cacheDir = new File(context.getApplicationContext().getCacheDir(), CACHE_DIR_NAME);
-        final Cache cache = new Cache(cacheDir, 20 * 1024 * 1024); // 20MB (toriaez)
-
-        final OkHttpClient client = new OkHttpClient();
-        client.setCookieHandler(createCookieHandler());
-        client.setSslSocketFactory(sslContext.getSocketFactory()); // XXX
-        client.setHostnameVerifier(new NonVerifyingHostnameVerifier()); // XXX
-        client.setCache(cache);
-        return client;
-    }
-
-    private static class TrustEveryoneX509TrustManager implements X509TrustManager {
-        @Override
-        public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-            throw new CertificateException();
-        }
-
-        @Override
-        public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-            // Trust any certificate!!!
-        }
-
-        @Override
-        public X509Certificate[] getAcceptedIssuers() {
-            return new X509Certificate[0];
-        }
-    }
-
-    private static class NonVerifyingHostnameVerifier implements HostnameVerifier {
-        @Override
-        public boolean verify(String hostname, SSLSession session) {
-            return true;
-        }
     }
 
     static class AuthenticationInterceptor implements RequestInterceptor {
